Changed type.

+++ src/main/java/nl/esciencecenter/wordembedding/commandline/Word2VecPMIComparisonCommandLineArguments.java
@@ -15,7 +15,7 @@ public class Word2VecPMIComparisonCommandLineArguments
     @Parameter(names = {"-sampling_rate"}, description = "Probability that a word is removed from the vocabulary.")
     private int samplingRate = 0;
     @Parameter(names = {"-max_pairs"}, description = "Maximum number of pairs.")
-    private int maxPairs = 0;
+    private long maxPairs = 0;
     // Files
     @Parameter(names = {"-vocabulary_file"}, description = "Word2Vec vocabulary file.", required = true)
     private String vocabularyFilename = "";
@@ -56,7 +56,7 @@ public class Word2VecPMIComparisonCommandLineArguments
         return samplingRate;
     }
 
-    public int getMaxPairs()
+    public long getMaxPairs()
     {
         return maxPairs;
     }
